Revert "Fix Array bridges error && update swift3 syntax"
This reverts commit 0ec129014172e7f262e2d9d96888ae9745a3970b.

+++ UberSignalsTests/SwiftTests.swift
@@ -109,7 +109,7 @@ class SwiftTests: XCTestCase {
         }
 
         observer.onSwiftSignal.fire()("1")
-        observer.onSwiftDoubleSignal.fire()("1", NSNumber(value: 2))
+        observer.onSwiftDoubleSignal.fire()("1", NSNumber(value: 2 as Int))
         XCTAssertEqual(observed, 2, "Should have fired callback")
     }
 }
